Simplify dataset definition
The main change here is to take out looping through the years, there is another framework in ehrQL, measures, that does this for us. I also switched the naming of patient attributes to have the prefix has (instead of had), we often use this naming convention for our studies. To shorten the code I also removed some of the variables that were only used to define the population and explicitly mention them only inside the define_population function.