Modularising code and adding measures framework (#3)
* Take reading codelists out of dataset definition

Creating a separate module will make it easier to share this code across different dataset definitions that we might be working on for this project. This also fixes the path to themethylphenidate_codelist, before the path was pointing to the same NHSD Primary Care Refset for ADHD.

* Simplify dataset definition

The main change here is to take out looping through the years, there is another framework in ehrQL, measures, that does this for us. I also switched the naming of patient attributes to have the prefix has (instead of had), we often use this naming convention for our studies. To shorten the code I also removed some of the variables that were only used to define the population and explicitly mention them only inside the define_population function.

* Add measures definition calculating yearly ADHD prevalence

This is using the ehrQL measures framework to run the same query for multiple time intervals. I tried to match your original 7 years, although I'd recommend to reduce this to 3 years for the first run to keep the run times short.

* Remove `report.py` for now as it may no longer be needed